Move remaining osu-specific implementations to OsuLegacySkin
